queue: chunk & iterator have been moved out to separate source file
