+++ src/queue/chunk.cpp
@@ -5,8 +5,6 @@ extern std::shared_ptr<cocaine::framework::logger_t> grape_queue_module_get_logg
 #define LOG_ERROR(...) COCAINE_LOG_ERROR(grape_queue_module_get_logger(), __VA_ARGS__)
 #define LOG_DEBUG(...) COCAINE_LOG_DEBUG(grape_queue_module_get_logger(), __VA_ARGS__)
 
-const int STATE_ACKED = 1;
-
 ioremap::grape::chunk_meta::chunk_meta(int max)
 	: m_ptr(NULL)
 {
@@ -81,7 +79,7 @@ bool ioremap::grape::chunk_meta::ack(int32_t pos, int state)
 	}
 
 	chunk_entry &entry = m_ptr->entries[pos];
-	if (entry.state != state && state == STATE_ACKED) {
+	if (entry.state != state && state == chunk_entry::STATE_ACKED) {
 		m_ptr->acked++;
 	} else {
 		LOG_INFO("\tmeta.ack: pos: %d, was already acked", pos, entry.state);
@@ -451,7 +449,7 @@ bool ioremap::grape::chunk::push(const ioremap::elliptics::data_pointer &d)
 bool ioremap::grape::chunk::ack(int pos, bool write)
 {
 	//FIXME: check if pos < low < high
-	m_meta.ack(pos, STATE_ACKED);
+	m_meta.ack(pos, chunk_entry::STATE_ACKED);
 	if (write) {
 		write_meta();
 	}
